requirements.txt
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
ftfy
pillow
regex
torch
torchvision
tqdm
fastapi[standard]
